+++ lib/bloc_after_effect.dart
@@ -1,3 +1,4 @@
+export 'src/bloc_effect_builder.dart';
 export 'src/bloc_effect_consumer.dart';
 export 'src/bloc_effect_listener.dart';
 export 'src/effect_bloc.dart';